requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
flask-marshmallow==0.15.0
Flask-JWT-Extended==4.5.2
Flask-Cors==3.0.10
apifairy==1.3.0